requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.7.2
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.7
deepgram-sdk==3.2.3
Django==4.2.9
django-cors-headers==4.3.1
django-currentuser==0.6.1
django-environ==0.11.2
django-filter==23.5
django-jazzmin==2.6.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
google-auth==2.26.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.1
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydub==0.25.1
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
verboselogs==1.7
websockets==12.0
yarl==1.9.4
zipp==3.17.0
stripe==12.0.0
